+++ feincms3_cookiecontrol/models.py
@@ -135,6 +135,12 @@ class CookieScript(models.Model):
         if (stripped := self.inject_else.strip()) and stripped[0] != "<":
             errors["inject_else"] = msg
 
+        msg = gettext("Entering <noscript> tags doesn't make sense.")
+        if "<noscript" in self.inject_if:
+            errors["inject_if"] = msg
+        if "<noscript" in self.inject_else:
+            errors["inject_else"] = msg
+
         if errors:
             raise ValidationError(errors)